requirements.txt
streamlit~=1.19.0
Pillow~=9.4.0
sqlmodel~=0.0.8
pandas~=1.5.3
bcrypt~=4.0.1
requests~=2.28.2